+++ lib/src/io.dart
@@ -2,10 +2,10 @@
 library where.io;
 
 import 'dart:async';
-import 'dart:io';
+import 'dart:io' as io;
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
 
-part 'io/file_stats.dart';
+part 'io/file_stat.dart';
 part 'io/finder.dart';
 part 'io/where.dart';