+++ test/all.dart
@@ -1,10 +1,10 @@
-import 'file_stats_test.dart' as file_stats_test;
+import 'file_stat_test.dart' as file_stat_test;
 import 'finder_test.dart' as finder_test;
 import 'where_test.dart' as where_test;
 
 /// Tests all the features of the package.
 void main() {
-  file_stats_test.main();
+  file_stat_test.main();
   finder_test.main();
   where_test.main();
 }